replace broken/obsolete delete-workspaces tool

+++ api/src/main/java/org/pmiops/workbench/impersonation/ImpersonatedWorkspaceServiceImpl.java
@@ -129,7 +129,8 @@ public class ImpersonatedWorkspaceServiceImpl implements ImpersonatedWorkspaceSe
   }
 
   @Override
-  public void deleteWorkspace(String username, String wsNamespace, String wsId) {
+  public void deleteWorkspace(
+      String username, String wsNamespace, String wsId, boolean deleteBillingProjects) {
     final DbUser dbUser = userDao.findUserByUsername(username);
 
     // also confirms that the workspace exists in the DB
@@ -162,21 +163,27 @@ public class ImpersonatedWorkspaceServiceImpl implements ImpersonatedWorkspaceSe
     workspaceDao.saveWithLastModified(
         dbWorkspace.setWorkspaceActiveStatusEnum(WorkspaceActiveStatus.DELETED), dbUser);
 
-    try {
-      // use the real FirecloudService here because impersonation is not needed;
-      // billing projects are owned by the App SA
-      firecloudService.deleteBillingProject(wsNamespace);
-      billingProjectAuditor.fireDeleteAction(wsNamespace);
-    } catch (Exception e) {
-      String msg =
-          String.format("Error deleting billing project %s: %s", wsNamespace, e.getMessage());
-      logger.warning(msg);
+    if (deleteBillingProjects) {
+      try {
+        // use the real FirecloudService here because impersonation is not needed;
+        // billing projects are owned by the App SA
+        firecloudService.deleteBillingProject(wsNamespace);
+        billingProjectAuditor.fireDeleteAction(wsNamespace);
+      } catch (Exception e) {
+        String msg =
+            String.format("Error deleting billing project %s: %s", wsNamespace, e.getMessage());
+        logger.warning(msg);
+      }
     }
   }
 
   @Override
   public void deleteOrphanedRawlsWorkspace(
-      String username, String wsNamespace, String googleProject, String wsId) {
+      String username,
+      String wsNamespace,
+      String googleProject,
+      String wsId,
+      boolean deleteBillingProjects) {
     final DbUser dbUser = userDao.findUserByUsername(username);
 
     try {
@@ -197,24 +204,16 @@ public class ImpersonatedWorkspaceServiceImpl implements ImpersonatedWorkspaceSe
       throw new ServerErrorException(e);
     }
 
-    // TODO
-    // I can't figure out how to make this work from a tool
-
-    // 2023-05-03 18:52:06.815  WARN 20371 --- [           main]
-    // o.p.w.i.ImpersonatedWorkspaceServiceImpl :
-    // Error deleting billing project aou-rw-test-cce862eb: No qualifying bean of type
-    // 'org.pmiops.workbench.rawls.api.BillingV2Api' available: expected at least 1 bean which
-    // qualifies as autowire candidate. Dependency annotations:
-    // {@org.springframework.beans.factory.annotation.Qualifier(value="serviceAccountBillingV2Api")}
-
-    //    try {
-    //      // use the real FirecloudService here because impersonation is not needed;
-    //      // billing projects are owned by the App SA
-    //      firecloudService.deleteBillingProject(wsNamespace);
-    //    } catch (Exception e) {
-    //      String msg =
-    //          String.format("Error deleting billing project %s: %s", wsNamespace, e.getMessage());
-    //      logger.warning(msg);
-    //    }
+    if (deleteBillingProjects) {
+      try {
+        // use the real FirecloudService here because impersonation is not needed;
+        // billing projects are owned by the App SA
+        firecloudService.deleteBillingProject(wsNamespace);
+      } catch (Exception e) {
+        String msg =
+            String.format("Error deleting billing project %s: %s", wsNamespace, e.getMessage());
+        logger.warning(msg);
+      }
+    }
   }
 }
